backend/ecommercewebsite/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2019.6.16
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==1.2.0
Django==4.1.7
django-allauth==0.53.1
django-filter==2.2.0
django-heroku==0.3.1
django-location-field==2.7.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-link-header-pagination==0.1.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extensions==0.7.1
drf-flex-fields==1.0.2
environs==9.5.0
et-xmlfile==1.1.0
Faker==13.14.0
filelock==3.7.0
gunicorn==20.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
joblib==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.19.0
MouseInfo==0.1.3
numpy==1.22.4
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
pandas==1.4.2
pendulum==2.1.2
Pillow==9.4.0
pimsampas==0.3
platformdirs==2.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyperclip==1.8.2
pypi-install==0.0.5
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytweening==1.0.4
pytz==2019.2
pytzdata==2020.1
requests==2.22.0
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
soupsieve==2.3.2.post1
sqlparse==0.3.0
structlog==22.3.0
swapper==1.3.0
threadpoolctl==3.1.0
typing-extensions==3.10.0.0
tzdata==2022.1
unclechat==0.0.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==20.14.1
whitenoise==6.4.0
wikipedia==1.4.0
